% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{convert_ts}
\alias{convert_ts}
\title{Converts ts timeries to tsibble format.}
\usage{
convert_ts(ts_data, pivot_longer = TRUE)
}
\arguments{
\item{pivot_longer}{Boolean. If TRUE assumes that each column is a different
unit of observation. If FALSE assumes that each column is a different
variable.}
}
\value{
tsibble object
}
\description{
Converts ts timeries to tsibble format.
}
