% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{look_at_ts}
\alias{look_at_ts}
\title{Visualises a time series.}
\usage{
look_at_ts(data, pivot_longer = TRUE)
}
\arguments{
\item{pivot_longer}{Logical. If TRUE each column is an observation unit or
key, if FALSE each column is taken to be a variable. Does not apply if data
is already a tsibble.}
}
\description{
Visualises a time series.
}
