% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tsibble_colinfo}
\alias{tsibble_colinfo}
\title{Computes information about columns of a tsibble object.}
\usage{
tsibble_colinfo(tsdata)
}
\arguments{
\item{tsdata}{A tsibble object.}
}
\value{
tibble data frame containing metadata about each column:
     is numeric, is measured, is key and is time index.
}
\description{
Computes information about columns of a tsibble object.
}
