% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tsibble_to_long}
\alias{tsibble_to_long}
\title{Transforms a tsibble to long format tibble.}
\usage{
tsibble_to_long(tsdata)
}
\value{
Should output a tsibble of the format index, key, variable, value.
 If there is only one key, then it outputs index, variable, value. 
 If there is only one variable then it outputs index, key, variable.
 If there is only one key and one variable, then it outputs index, variable.
}
\description{
Transforms a tsibble to long format tibble.
}
