% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{versus_plot}
\alias{versus_plot}
\title{Plots all variables against one xvar.}
\usage{
versus_plot(tsdata, xvar)
}
\arguments{
\item{tsdata}{(tsibble) Data to be plotted, tsibble time series object, 0 or one keys at least 2 measured variables.}

\item{xvar}{(string) Variable to put on x axis.}
}
\description{
Plots all variables against one xvar.
}
