% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{versus_plot}
\alias{versus_plot}
\title{Plots all variables against one xvar.}
\usage{
versus_plot(tsdata, xvar)
}
\arguments{
\item{xvar}{Variable to put on x axis.}
}
\description{
Plots all variables against one xvar.
}
